Add intermediary output
