+++ buildSrc/src/main/java/quilt/internal/MappingsPlugin.java
@@ -3,24 +3,11 @@ package quilt.internal;
 import org.gradle.api.Plugin;
 import org.gradle.api.Project;
 import org.gradle.api.tasks.TaskContainer;
-import quilt.internal.tasks.build.BuildMappingsTinyTask;
-import quilt.internal.tasks.build.CompressTinyTask;
-import quilt.internal.tasks.build.DropInvalidMappingsTask;
-import quilt.internal.tasks.build.GeneratePackageInfoMappingsTask;
-import quilt.internal.tasks.build.InvertPerVersionMappingsTask;
-import quilt.internal.tasks.build.MergeTinyTask;
-import quilt.internal.tasks.build.MergeTinyV2Task;
-import quilt.internal.tasks.build.TinyJarTask;
+import quilt.internal.tasks.build.*;
 import quilt.internal.tasks.jarmapping.MapNamedJarTask;
 import quilt.internal.tasks.jarmapping.MapPerVersionMappingsJarTask;
 import quilt.internal.tasks.lint.MappingLintTask;
-import quilt.internal.tasks.setup.DownloadMinecraftJarsTask;
-import quilt.internal.tasks.setup.DownloadMinecraftLibrariesTask;
-import quilt.internal.tasks.setup.DownloadPerVersionMappingsTask;
-import quilt.internal.tasks.setup.DownloadVersionsManifestTask;
-import quilt.internal.tasks.setup.DownloadWantedVersionManifestTask;
-import quilt.internal.tasks.setup.ExtractServerJarTask;
-import quilt.internal.tasks.setup.MergeJarsTask;
+import quilt.internal.tasks.setup.*;
 import quilt.internal.tasks.unpick.CombineUnpickDefinitionsTask;
 import quilt.internal.tasks.unpick.RemapUnpickDefinitionsTask;
 import quilt.internal.tasks.unpick.gen.OpenGlConstantUnpickGenerator;
@@ -56,6 +43,10 @@ public class MappingsPlugin implements Plugin<Project> {
 
         tasks.create(GeneratePackageInfoMappingsTask.TASK_NAME, GeneratePackageInfoMappingsTask.class);
         tasks.create(MappingLintTask.TASK_NAME, MappingLintTask.class);
+
+        tasks.create(DownloadIntermediaryMappingsTask.TASK_NAME, DownloadIntermediaryMappingsTask.class);
+        tasks.create(InvertIntermediaryMappingsTask.TASK_NAME, InvertIntermediaryMappingsTask.class);
+        tasks.create(MergeIntermediaryTask.TASK_NAME, MergeIntermediaryTask.class);
     }
 
     public static MappingsExtension getExtension(Project project) {
