+++ buildSrc/src/main/java/quilt/internal/tasks/setup/DownloadIntermediaryMappingsTask.java
@@ -1,0 +1,9 @@
+package quilt.internal.tasks.setup;
+
+public class DownloadIntermediaryMappingsTask extends AbstractDownloadMappingsTask {
+    public static final String TASK_NAME = "downloadIntermediaryMappings";
+
+    public DownloadIntermediaryMappingsTask() {
+        super("intermediary");
+    }
+}
